requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
aiohttp==3.8.5
Pillow==10.0.0
python-magic==0.4.27
APScheduler==3.10.4
PyJWT==2.8.0
werkzeug==2.3.7
flask-limiter==3.5.0
redis==5.0.1

